# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CRM Cia da Vacina API
  version: 0.1.0
  description: |
    Contrato Sprint 1+ — Auth, Users, Units, Me.
    Endpoints de Inbox/WhatsApp entram nas sprints seguintes mantendo este estilo.
servers:
  - url: http://localhost:8080/api/v1
    description: Local API
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Novo par de tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users:
    get:
      tags: [Users]
      summary: Listar usuários
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Users]
      summary: Criar usuário
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Obter usuário
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      summary: Atualizar usuário
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{userId}/units:
    put:
      tags: [Users]
      summary: Substituir vínculo do usuário com unidades
      operationId: setUserUnits
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [unit_ids]
              properties:
                unit_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '204':
          description: Atualizado
        '403':
          $ref: '#/components/responses/Forbidden'
  /units:
    get:
      tags: [Units]
      summary: Listar unidades
      operationId: listUnits
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Unit'
    post:
      tags: [Units]
      summary: Criar unidade
      operationId: createUnit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnitRequest'
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unit'
        '403':
          $ref: '#/components/responses/Forbidden'
  /units/{unitId}:
    get:
      tags: [Units]
      summary: Obter unidade
      operationId: getUnit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UnitId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unit'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Units]
      summary: Atualizar unidade
      operationId: updateUnit
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UnitId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUnitRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unit'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UnitId:
      name: unitId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
  responses:
    BadRequest:
      description: Validação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: unauthorized
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    LoginResponse:
      type: object
      required: [access_token, refresh_token, expires_in, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Segundos até expirar o access_token
        user:
          $ref: '#/components/schemas/User'
    UserRole:
      type: string
      enum: [admin, manager, supervisor, agent]
...
    User:
      type: object
      required: [id, email, name, role, active]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
        unit_ids:
          type: array
          items:
            type: string
            format: uuid
...
    CreateUserRequest:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        unit_ids:
          type: array
          items:
            type: string
            format: uuid
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
        password:
          type: string
          minLength: 8
    Unit:
      type: object
      required: [id, name, code, timezone, active]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        timezone:
          type: string
          example: America/Sao_Paulo
        active:
          type: boolean
    CreateUnitRequest:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        timezone:
          type: string
          default: America/Sao_Paulo
    UpdateUnitRequest:
      type: object
      properties:
        name:
          type: string
        timezone:
          type: string
        active:
          type: boolean
...